🔧 (main.yml): add GitHub context variables and temporarily disable build jobs
Add GitHub context variables to the workflow to enhance the ability to access repository and reference information dynamically. This change is crucial for workflows that need to differentiate between push and pull request events and access specific repository details accordingly.

Temporarily disable the "Build, Test, Sonar Cloud Analysis, & Package" and "Build Documentation output" jobs by setting their conditions to false. This is likely done to prevent these jobs from running during a specific phase of development or testing, possibly to save resources or focus on other parts of the workflow.