📝 (default.html): add giscus backlink meta tag for improved comment integration
Add a meta tag for giscus backlink to enhance the integration with
the giscus comment system. This change helps in associating comments
with the correct page URL, improving the user experience and
functionality of the comment system. Additionally, minor formatting
adjustments are made to the Google Analytics script for better
readability.