Fix #2521 reported by @nirkal \n\n ✨ (ServiceCollectionExtensions.cs, TfsGitRepositoryTool.cs): add TfsChangeSetMappingTool to service collection and ensure required service retrieval
Add `TfsChangeSetMappingTool` to the service collection to enable its configuration and usage within the application. Change the retrieval method of `TfsChangeSetMappingTool` from `GetService` to `GetRequiredService` to ensure that the service is available and to prevent potential null reference exceptions. This change enhances the robustness and configurability of the application by ensuring necessary services are properly registered and retrieved.