Preserve order of mappings in TfsNodeStructureToolOptions
The mappings were defined as dictionary, which does not have an order. The items added to dictionary are returned in random order. In case of mappings, we need to preserver the order of them, because we need them to be matched in order specified in configuration. This is important, because more than one mapping can match input value.